Minimal working version finished. Adding index chacks and storing CelebA next.

+++ util.py
@@ -3,6 +3,7 @@ import torch
 import math
 
 def writeManifest(dirname, totalRecords, batchSize, numBatches):
+    """Writes the correctly formatted manifest."""
     f = open(os.path.join(dirname, 'manifest'), 'w')
     s = "This directory is formatted for a BatchDataset.\n"
     s += "This file contains metadata.\n\n"
@@ -15,7 +16,13 @@ def writeManifest(dirname, totalRecords, batchSize, numBatches):
 
 
 def parseManifest(dirname):
-    f = open(os.path.join(dirname, 'manifest'), 'r') # Put this in a try-except
+    """Parses the manifest file into totalRecords, batchSize, and numBatches. 
+If there is no manifest file, will throw an error; bad formatting will lead to 
+undefined behavior."""
+    try:
+        f = open(os.path.join(dirname, 'manifest'), 'r')
+    except FileNotFoundError:
+        raise OSError("Directory '" + dirname + "' has no manifest, is not a formatted directory.")
     s = f.read()
     f.close()
     l = s.split('\n')[3:]
@@ -25,13 +32,48 @@ def parseManifest(dirname):
     return totalRecords, batchSize, numBatches
 
 
+def isDirectoryFormatted(dirname):
+    """Checks that only the manifest and the correct number of batch files are present. 
+However, it does not verify the contents of the batchs; storing incorrect batches leads 
+to undefined behavior."""
+    try:
+        _, _, numBatches = parseManifest(dirname)
+    except OSError:
+        return False
+
+    l = os.listdir(dirname)
+    l.remove('manifest') # Only batches should remain
+
+    # Fastest check: check length
+    if len(l) != numBatches:
+        return False
+
+    # Check that each name begins with 'batch'
+    for fname in l:
+        if fname[:5] != 'batch':
+            return False
+    
+    # Check that all the suffixes are correct
+    inds = [int(fname[5:]) for fname in l]
+    inds.sort()
+    for i in range(numBatches):
+        if inds[i] != i:
+            return False
+
+    # We passed all checks.
+    return True
+
+
 def verifyDirEmpty(dirname, cleanDir=False):
+    """Checks if directory is empty. 
+If not, and cleanDir is set to True, empties the directory 
+and still returns True."""
     if len(os.listdir(dirname)) > 0:
         if cleanDir:
             for fname in os.listdir(dirname):
                 os.remove(os.path.join(dirname, fname))
         else:
-            raise OSError("Directory '" + dirname  + "' not empty; choose a new directory or set cleanDir=True.")
+            return False
     return True 
 
 
@@ -42,8 +84,12 @@ then used by BatchDataset. This only needs to be run once, so it's a utility.
 Adjust batchSize so two batch leaves enough (CPU) RAM for normal processes, 
 but is as large as possible. BathDataset sometimes loads two batches, 
 to avoid waiting on a disk read."""
-    if verifyDirEmpty(dirname, cleanDir=cleanDir) and verbose:
-        print("Directory '" + dirname + "' is empty.")
+    if verifyDirEmpty(dirname, cleanDir=cleanDir):
+        if verbose:
+            print("Directory '" + dirname + "' is empty.")
+    else:
+        raise OSError("Directory '" + dirname  + "' not empty; choose a new directory or set cleanDir=True.")
+
     totalRecords = len(source)
     numBatches = math.ceil(totalRecords / batchSize)
     if writeManifest(dirname, totalRecords, batchSize, numBatches) and verbose:
@@ -57,7 +103,7 @@ to avoid waiting on a disk read."""
             print("Records up to " + str(recordId + 1) +  " saved; " + str(batchId + 1) + " batches written.")
 
     # Main loop to store all the records.
-    # By default, python list are the batches, but the BatchDataset class will work with any container.
+    # By default, python list are the batches, but the BatchDataset class will work with any indexable container.
     res = []
     for i in range(totalRecords):
         res.append(source[i])
